+++ libs/libxx/Makefile
@@ -35,8 +35,6 @@ ifeq ($(CONFIG_UCLIBCXX),y)
 include uClibc++/Make.defs
 else ifeq ($(CONFIG_LIBCXX),y)
 include libcxx/Make.defs
-else ifeq ($(CONFIG_LIBCXXMINI),y)
-include libcxxmini/Make.defs
 endif
 
 ifeq ($(CONFIG_ETL),y)
@@ -45,6 +43,8 @@ endif
 
 ifeq ($(CONFIG_LIBCXXABI),y)
 include libcxxabi/Make.defs
+else ifeq ($(CONFIG_LIBMINIABI),y)
+include libminiabi/Make.defs
 endif
 
 # Object Files
