+++ libs/libxx/libminiabi/libxx_newa.cxx
@@ -1,5 +1,5 @@
 //***************************************************************************
-// libs/libxx/libcxxmini/libxx_new.cxx
+// libs/libxx/libminiabi/libxx_newa.cxx
 //
 // SPDX-License-Identifier: Apache-2.0
 //
@@ -31,6 +31,14 @@
 #include <nuttx/lib/lib.h>
 
 //***************************************************************************
+// Pre-processor Definitions
+//***************************************************************************
+
+//***************************************************************************
+// Private Data
+//***************************************************************************
+
+//***************************************************************************
 // Operators
 //***************************************************************************
 
@@ -48,7 +56,7 @@
 //
 //***************************************************************************
 
-FAR void *operator new(std::size_t nbytes)
+FAR void *operator new[](std::size_t nbytes)
 {
   // Perform the allocation
 
@@ -71,7 +79,7 @@ FAR void *operator new(std::size_t nbytes)
   return alloc;
 }
 
-FAR void *operator new(std::size_t nbytes, FAR void *ptr)
+FAR void *operator new[](std::size_t nbytes, FAR void *ptr)
 {
   DEBUGASSERT(ptr != NULL);
